refactor: clean up select type option logic
