+++ frontend/appflowy_flutter/lib/plugins/database/grid/presentation/widgets/header/type_option/select/select_option_editor.dart
@@ -43,7 +43,9 @@ class SelectOptionTypeOptionEditor extends StatelessWidget {
           BlocListener<EditSelectOptionBloc, EditSelectOptionState>(
             listenWhen: (p, c) => p.deleted != c.deleted,
             listener: (context, state) {
-              state.deleted.fold(() => null, (_) => onDeleted());
+              if (state.deleted) {
+                onDeleted();
+              }
             },
           ),
           BlocListener<EditSelectOptionBloc, EditSelectOptionState>(
